TLBC2App: implement exposure time setting.
Both set and get methods are used to ensure that the readback value is
consistent with the hardware. Note that the superclass could do both set
the double parameter and call callbacks, but this is done here to update
the readback according to the hardware feedback.

The unit used in the manufacturer library is milliseconds. Convert it to
seconds so that it is in SI units. Convertion is done in the record
level, so that tooling from EPICS base can be properly used. This
simplifies the driver, as unit convertion does not to be handled
explicitly. ASLO field has been used instead of LINR and ESLO, as they
seem not to work out of the box. Further investigation is required to
understand and improve this.

As ADCore defines precision to be only 3 digits and the unit is
milliseconds, the resolution becomes poor for this define. Thus,
increase it by 3 decimals places to include the 1e-3 factor. Default
VAL field is also changed to 0.1, so that it is in the valid range for
BC210CU camera, which is limited to 500ms.