TLBC2App: parametrize device reset on connect.
When connecting to the device, we always reset the device. However, if
the IOC is to be used along with pre-defined values, we cannot do that.
Make this an option when instantiating the driver, so that it is up to
the user define whether the define should be reset or not.