TLBC2App: implement gain setting.
Gain is checked to be in the valid range directly in the driver code for
simplicity. This should be refactored to properly fill DRVL and DRVH
fields of the Gain process variable. Doing so probably means adding
custom records to retrieve the value from the driver and forward them to
these fields.