TLBC2App: use object-oriented parameter access.
Logic for handling exposure time and gain are pretty much the same,
since the TLBC2 library is consistent with its API. Other simple
parameters seems to follow the same rationale, which would generate yet
more duplicated logic changing only library function calls. Make an
object-oriented manner to call the library functions, which enables to
reuse the code easily.

Adding a new parameter with this design simply means adding a new entry
to the parameter mapping. Since not all parameters support querying the
valid range of values, this is made optional to specify. If it is
defined, it will be used when setting a value.

The range check should be actually done in the EPICS record level by
correctly defining DRVL and DRVH fields in the corresponding records.
This is not done here for simplicity.