iocs: make st.cmd a template.
Configure driver and plugin parameterized, so that it can be reused
instead of replicated. This way, st.cmd is a template to be copied and
changed with the relevant parameters without ever bothering to redefine
everything again.

This also makes the st.cmd shorter, cleaner, and direct to the relevant
definitions.