TLBC2App: readback even when failing to set point.
Readback is a side-effect of writing a value to the driver which should
be unconditional in its success. Make sure the camera is queried and the
corresponding readback value is set even when the provided value is
invalid (outside the specified range) or the command fails for an
unexpected reason.