TLBC2App: put scan calculations as attributes in NDArray.
`TLBC2_get_scan_data` computes statistics based on the just acquired
image. Export these calculations as NDAttributes in the outcoming
NDArray, so that file plugins that follows, such as NDFileHDF5 plugin,
can handle these metadata atomically with the data.

All statistics have been exported, except 1-dimensional profile arrays
and basic statistics, including

- minimum and maximum intensity;
- saturation;
- standard deviation in x, y and radial.

They may be recomputed in higher level if required.

Descriptions are based on the comments provided in the
TLBC1_Calculations.h header file. Refer to it to further details.

Static asserts are done to ensure that the casting done internally in
ADCore is really what we are assuming it to be.