TLBC2App: implement auto exposure parameter setting.
Auto exposure feature allows to use the TLBC2 library provided
heuristics for setting the exposure time automatically based on the
acquired image. Export this feature to EPICS-layer, so that users can
choose on whether to use it other than AcquireTime.

As the handling for setting and reading back for real, integer, and
boolean values are the same except for the type itself, extract this it
to a templated method.

Auto exposure is reset to zero after the user sets a exposure time
manually. This should be improved to handle the reading from the library
instead, so that is indeed synchronized with it. For simplicity, this is
assumed to be always true here.