TLBC2App: reset calc area before setting ROI.
When we set a ROI to a region larger than it was before, the library
violates access (segfaults) at `TLBC1_fillCalculationAreaMask()`. This
is probably because the mask buffer is smaller than it should to fill
the new ROI.

Workaround this by defining the calculation area larger before setting
the ROI, and then rolling back to the previous configuration. Maximum
size is set to make the logic simpler, as this way any ROI will be valid
regardless of its size. Center is set to zero, as we only care about the
buffer size.