TLBC2App: implement ROI definition parameters.
ROI is very useful for BC210CU device, since its sensor is rather large,
while the beam may hit a really small area of the sensor. Add support
for it based on the ADDriver-defined parameters.

Since the ROI not always can be properly setup at start-up, due to the
order of the values to be configured, let's make its PINI to NO. This
ensures that the IOC does not start in an intermediate (invalid) state.