TLBC2App: implement clip level for auto calc area.
Automatic calculation area can be defined based on a parameter called
clip level. Not using this makes the full size ROI be defined, which
significantly influences the beam width statistics, depending on the ROI
size. Make this available through the IOC, so that it can be ensure to
be correct upon data collection.

The value of 1% is used as default here, as it is also the default in
the ThorLabs Beam application.