iocs: use autosave module in IOC.
A independent cmd file is used for loading the configuration, so that it
can be reused or substituted when written a specific st.cmd file. As
loading the configuration must be done after the iocInit(), but defining
them must be done before, use another cmd script to do the
post-initializaton setup.

Plugins are left in a separate file, as they should match the
instantiated plugins in plugins.cmd file. This way, it is easier to
match one configration file with the other.

Autosave does not need to be defined as a dependency in the IOC's
src/Makefile, as commonDriverMakefile automatically adds it if
"AUTOSAVE" is defined. Therefore, simply defining it in areaDetector
global configuration file (or RELEASE.local) is sufficient to be
possible to use the st.cmd template as defined. Not defining it implies
the user must remove the cmd loading from the template.

Calc is defined in the search path, since its sseq record is used by
NDPluginStats, which is also configured to be autosaved.