TLBC2App: implement ambient light correction setting.
Ambient light correction can only be enabled if it has already been
previously run and has a volatile metadata available. Make two PVs for
handling this: one for computing the metadata when the user guarantees
there is no beam in the sensor, and other for enabling it. This allows
the user to make a conscious decision about when to compute the
correction and when to apply it.

When attempting to enable before computing, the vendor library correctly
reports and error. This makes the readback value to keep at disabled,
which is what we want. Thus, no further checking is required to
guarantee this consistency.

A custom enum is defined for possible values from correction status, as
the provided headers does not define them. The documented return values
have been used to define them.