TLBC2App: implement clip level setting.
Clip level is used for computing the BeamWidthClip{X,Y} statistics. Such
value should be exposured for allowing defining the desired parameter
and also to eventually include it as a attribute in the resulting
dataset, along with its result.

The value 13.5% has been chosen the default, as it is the default in the
vendor software.