TLBC2App: implement temperature reading.
Temperature monitoring is important, since the device heats quite fast
and its recommended operation temperature is low. Thus, having this
exported to the control system eases making sure this is satisfied.

This is done through readFloat64, so that we can reuse the EPICS base
functionality for creating and managing scan threads. An alternative
would be to reimplement this in the driver as a poller and keeping I/O
Intr for the record [1]. Since we can override the scan value at IOC
initialization, use that to avoid replication.

Both TemperatureActual and Temperature_RBV are updated, to keep their
values valid. Note that, for the sake of simplicity, the temperature is
read twice. A refactoring should be considered to read this once
instead. Calling parameter callbacks is required to have the
Temperature_RBV updated, as it is an I/O Intr record.

[1]: https://epics.anl.gov/tech-talk/2018/msg02164.php