TLBC2App: expose beam width as a PV.
Beam width measured from the peak down to the clip level provide a
useful insight for operators regarding the beam caracteristics. Expose
this as a read-only PV for prompt feedback.

While this in theory could be taken from `TLBC2_get_beam_width`, it is
already available when a new scan data is processed. So just export it.