TLBC2App: expose saturation statistics as a PV.
The vendor library already computes the pixel saturation ratio, which is
the ratio of the maximum intensity in the scan to the dynamic range of
the sensor.

This is important to be tracked when configuring the camera for an
experiment without auto-exposure feature. One may base on this value to
properly configure Gain and ExposureTime.