iocs: use dimensions from BC210CU/M
This is the device used during tests. Change the dimensions to its
sensor size so that both MaxSizeX_RBV and MaxSizeY_RBV have properly
configured values at initialization.

An alternative way to have this working for all models is probably to
use the TLBC2_get_fullsize_image function, which returns the image
dimensions. However, this implies that at least one acquisition must be
done before the value is available.