TLBC2App: request new measurement before getting data.
`get_scan_data()` is documented to "read out" an image from the beam
profiler. However, based on tests, this is done only after the data is
analyzed, which makes the image off by one when correlating with other
process variables from an experimental setup and time.

Request a new data before each acquisition, so that we always retrieve a
just acquired image from the camera.

Fixes: e15b61a (TLBC2App: add acquire thread.)