TLBC2App: always set ambient light correction status.
The status can't be changed without the IOC triggering something,
because the ambient light correction happens in the library used by the
IOC, not the hardware; but we set the value unconditionally, to simplify
the code.

TLBC2_run_ambient_light_correction returns an error code if it fails, so
it's not necessary to re-check
TLBC2_get_ambient_light_correction_status.