TLBC2App: Remove ADTemperature
The ADDriver interface specifies 2 temperature PVs: ADTemperature and
ADTemperatureActual. Since ADTemperature is actually a pair of
read/write PVs, it should be used for controlling the device's
temperature, therefore ADTemperature is intended to be set to the
desired temperature. ADTemperatureActual is the actual temperature
inside the device, which can differ from ADTemperature.

Since the TLBC2 and TLBC1 libraries don't offer any control over the
device's temperature, there's no reason to set the ADTemperature PV, as
that could cause confusion if a user thinks they can control the
device's temperature.

Since we no longer need to do anything other than updating the readback
value when reading, we can make the logic more generic for any internal
float64 in the parameter library

Fixes: 207435c (TLBC2App: implement temperature reading., 2024-07-12)