docs: Add docs
Added .rst docs roughly following the structure of other AD driver docs
such as ADSimDetector [1] and ADmarCCD [2]. Most other drivers have a
table containing all the supported standard parameters from ADDriver.
Since we felt that this was redundant, the docs contain a list of all
the supported parameters along with a table which documents only the
standard parameters which have specific implementation details in the
driver.

For now the link in the README links directly to the .rst file, however
once ADTLBC2 is moved into the AD repo we can link to the generated
.html documentation.

[1]: https://areadetector.github.io/areaDetector/ADSimDetector/simDetector.html
[2]: https://areadetector.github.io/areaDetector/ADmarCCD/marCCDDoc.html