TLBC2App: Add multiple and continuous capture
Added support for the ImageMultiple and ImageContinuous mode. Moved the
core logic of acquiring exactly one image into acquire_image, and the
logic for doing one whole acquisition into do_acquisition, which returns
when the acquisition is done or if there was an error.

The try_wait_acquire_period function takes the time that the current
single-image acquisition started and returns when acquirePeriod is over
or stop_acquire_event has been triggered. If it has been triggered the
function returns true and we return from the do_acquisition function,
stopping the acquisition. Otherwise we continue inside the
do_acquisition loop. Note that this is only done for ImageMultiple and
ImageContinuous mode, since there's no point in stopping a single image
acquisition.

The 0.001s delay when waiting for the stop_acquire_event was tested
and found to reliably work. It's possible that setting an even lower
value may still work, but that seems unnecessary since it's already
really fast.

Error reporting will be added in later commits.

The code for this was inspired by [1] and [2].

[1]: https://github.com/areaDetector/ADRIXSCam/pull/5
[2]: https://github.com/areaDetector/ADmarCCD/blob/master/marCCDApp/src/marCCD.cpp