TLCB2App: Update ADStatusMessage to signal errors
Since the current error messages are very small, the 40 character limit
should be enough for now. Note that since we are only updating
ADStatusMessage with errors, it will either be empty or contain the last
error that ocurred.

There are some places where we need to make sure to call
callParamCallbacks and others where the caller will already have to do
so. The first catch block inside the writeROI function does not need
call callParamCallbacks since the code after it will already do so.

readAcquireTime and writeParam expect its callers to call
callParamCallbacks, so we don't do need to call it.