TLBC2App: Set ADAcquire alarm on acquisition failure
On acquisition failure, the ADAcquire alarm status and severity are set
to Comm (communication) and major, respectively. Alarms are reset at the
start of every acquisition, and a write of 0 to ADAcquire while it's
already 0 will also reset alarms.

After this change the only ocasion where a write to ADAcquire returns
error is a write to 1 while ADAcquire is already 1, in other words,
starting an acquisition while one is already taking place.