iocs: load autosave db from its correct location
Autosave database should not be read from a relative path, but rather
the absolute path specified by AUTOSAVE.

Alternatively, we could configure tlbc2IOC to install required database
from $AUTOSAVE, and specify the path relative to $TOP. However, that's
not done for other database dependencies, and it is not the practice in
other example IOCs. So, let's stick with conventional direct dependence
on the installed artifact for now.

Fixes: 2b55e00 (iocs: add missing load database save_restoreStatus.db, 2025-07-18)